+++ rbx_dom_weak/src/dom.rs
@@ -75,6 +75,33 @@ impl WeakDom {
         }
     }
 
+    /// Returns an iterator that goes through every descendant Instance of the
+    /// root referent.
+    ///
+    /// The descendants are guaranteed to be top-down such that children come
+    /// after their parents.
+    #[inline]
+    pub fn descendants(&self) -> WeakDomDescendants {
+        self.descendants_of(self.root_ref)
+    }
+
+    /// Returns an iterator that goes through the descendants of a particular
+    /// [`Ref`]. The passed `Ref` *must* be a part of this `WeakDom`.
+    ///
+    /// ## Panics
+    ///
+    /// Panics if `referent` is not a member of this DOM.
+    #[inline]
+    pub fn descendants_of(&self, referent: Ref) -> WeakDomDescendants {
+        if !self.instances.contains_key(&referent) {
+            panic!("the referent provided to `descendants_of` must be a part of the DOM")
+        }
+        WeakDomDescendants {
+            dom: self,
+            queue: [referent].into(),
+        }
+    }
+
     /// Insert a new instance into the DOM with the given parent. The parent is allowed to
     /// be the none Ref.
     ///
@@ -353,6 +380,29 @@ impl WeakDom {
         }
 
         instance
+    }
+}
+
+/// A struct for iterating through the descendants of an Instance in a
+/// [`WeakDom`].
+///
+/// See: [`WeakDom::descendants`] and [`WeakDom::descendants_of`].
+#[derive(Debug)]
+pub struct WeakDomDescendants<'a> {
+    dom: &'a WeakDom,
+    queue: VecDeque<Ref>,
+}
+
+impl<'a> Iterator for WeakDomDescendants<'a> {
+    type Item = &'a Instance;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let instance = self
+            .queue
+            .pop_front()
+            .and_then(|r| self.dom.get_by_ref(r))?;
+        self.queue.extend(instance.children());
+        Some(instance)
     }
 }
 
@@ -736,5 +786,29 @@ mod test {
         } else {
             panic!("UniqueId property must exist and contain a Variant::UniqueId")
         };
+    }
+
+    #[test]
+    fn descendants() {
+        let mut dom = WeakDom::new(InstanceBuilder::new("ROOT"));
+
+        let child_1 = dom.insert(dom.root_ref(), InstanceBuilder::new("Folder"));
+        let sibling_1 = dom.insert(child_1, InstanceBuilder::new("Folder"));
+        let child_2 = dom.insert(dom.root_ref(), InstanceBuilder::new("Folder"));
+        let sibling_2 = dom.insert(child_1, InstanceBuilder::new("Folder"));
+
+        let mut descendants = dom.descendants();
+        assert_eq!(descendants.next().unwrap().referent(), dom.root_ref());
+        assert_eq!(descendants.next().unwrap().referent(), child_1);
+        assert_eq!(descendants.next().unwrap().referent(), child_2);
+        assert_eq!(descendants.next().unwrap().referent(), sibling_1);
+        assert_eq!(descendants.next().unwrap().referent(), sibling_2);
+        assert!(descendants.next().is_none());
+
+        let mut descendants_2 = dom.descendants_of(child_1);
+        assert_eq!(descendants_2.next().unwrap().referent(), child_1);
+        assert_eq!(descendants_2.next().unwrap().referent(), sibling_1);
+        assert_eq!(descendants_2.next().unwrap().referent(), sibling_2);
+        assert!(descendants_2.next().is_none());
     }
 }